refactor(server): extract listen-target utilities from bootstrap.ts
The `ListenTarget` union, `parseListenString`, `resolveBoundListenTarget`,
`formatListenTarget`, and the `WINDOWS_DRIVE_RE` regex all live with
the daemon bootstrap entrypoint. They are pure string/HTTP utilities
with no dependency on the daemon lifecycle, MCP server, or websocket
plumbing in the same file.

Move them to `packages/server/src/utils/listen-target.ts` so they can
be unit-tested in isolation and reused by future daemons (e.g. a
relay-side control channel). `bootstrap.ts` and its smoke test get a
new import; behaviour and call sites are unchanged.

`service-proxy.ts` has its own private `resolveBoundListenTarget`
that is intentionally not deduplicated here — that file is a
sibling-but-isolated copy and unifying them is a separate change
(REVIEW.md P2-3).
