+++ packages/server/src/server/bootstrap.smoke.test.ts
@@ -6,7 +6,8 @@ import pino from "pino";
 import { afterEach, describe, expect, test, vi } from "vitest";
 import { WebSocket } from "ws";
 
-import { createPaseoDaemon, parseListenString, type PaseoDaemonConfig } from "./bootstrap.js";
+import { createPaseoDaemon, type PaseoDaemonConfig } from "./bootstrap.js";
+import { parseListenString } from "../utils/listen-target.js";
 import { loadConfig } from "./config.js";
 import { AgentManagerShuttingDownError } from "./agent/agent-manager.js";
 import { hashDaemonPassword } from "./auth.js";
