+++ packages/server/src/utils/listen-target.ts
@@ -1,0 +1,80 @@
+import type { Server } from "node:http";
+
+export type ListenTarget =
+  | { type: "tcp"; host: string; port: number }
+  | { type: "socket"; path: string }
+  | { type: "pipe"; path: string };
+
+// Matches a Windows drive-letter path like C:\ or D:\
+const WINDOWS_DRIVE_RE = /^[A-Za-z]:\\/;
+
+export function parseListenString(listen: string): ListenTarget {
+  // 1. Windows named pipes: \\.\pipe\... or pipe://...
+  if (listen.startsWith("\\\\.\\pipe\\") || listen.startsWith("pipe://")) {
+    return {
+      type: "pipe",
+      path: listen.startsWith("pipe://") ? listen.slice("pipe://".length) : listen,
+    };
+  }
+  // 2. Explicit unix:// prefix
+  if (listen.startsWith("unix://")) {
+    return { type: "socket", path: listen.slice(7) };
+  }
+  // 3. Reject Windows absolute drive paths — they are not Unix sockets
+  if (WINDOWS_DRIVE_RE.test(listen)) {
+    throw new Error(`Invalid listen string (Windows path is not a valid listen target): ${listen}`);
+  }
+  // 4. POSIX absolute path (/ or ~) — Unix socket
+  if (listen.startsWith("/") || listen.startsWith("~")) {
+    return { type: "socket", path: listen };
+  }
+  // 5. Pure numeric — TCP port on 127.0.0.1
+  const trimmed = listen.trim();
+  if (/^\d+$/.test(trimmed)) {
+    const port = parseInt(trimmed, 10);
+    return { type: "tcp", host: "127.0.0.1", port };
+  }
+  // 6. host:port — TCP
+  if (listen.includes(":")) {
+    const lastColonIdx = listen.lastIndexOf(":");
+    const host = listen.slice(0, lastColonIdx);
+    const portStr = listen.slice(lastColonIdx + 1);
+    const parsedPort = parseInt(portStr, 10);
+    if (!Number.isFinite(parsedPort)) {
+      throw new Error(`Invalid port in listen string: ${listen}`);
+    }
+    const cleanHost = host.startsWith("[") && host.endsWith("]") ? host.slice(1, -1) : host;
+    return { type: "tcp", host: cleanHost || "127.0.0.1", port: parsedPort };
+  }
+  throw new Error(`Invalid listen string: ${listen}`);
+}
+
+export function resolveBoundListenTarget(
+  listenTarget: ListenTarget,
+  httpServer: Server,
+): ListenTarget {
+  if (listenTarget.type !== "tcp") {
+    return listenTarget;
+  }
+
+  const address = httpServer.address();
+  if (!address || typeof address === "string") {
+    throw new Error("HTTP server did not expose a TCP address after listening");
+  }
+
+  return {
+    type: "tcp",
+    host: listenTarget.host,
+    port: address.port,
+  };
+}
+
+export function formatListenTarget(listenTarget: ListenTarget | null): string | null {
+  if (!listenTarget) {
+    return null;
+  }
+  if (listenTarget.type === "tcp") {
+    return `${listenTarget.host}:${listenTarget.port}`;
+  }
+  return listenTarget.path;
+}
